rename file with wrong name
